feat: add prelimenary LvimUpdate command
